Add tools to assess birdnet performance
There is a bunch of labeled audio data from 2017/2018 from
the BUOW team, and these tools are required to assess the
performance of a labeled audio chunk that has been run through
birdnet. aggregate_audio_analysis.py was created by Nidhi and groups
all birdnet output in the same csv, as the audio chunk had to be
split into a smaller file to be assessed by birdnet. Then
normalize_birdnet_output.py takes in that file and changes the
output to include the 'no' for all the 3 second chunk
times that there were no detections. Then if you have the name of
that audio file, you input that and the scored data csv into
normalize_scored_output.py, to make the output the same format
at the birdnet normalized output so they can be directly compared.
Then you input those file paths into assess_performance.py and it
will output the accuracy, precision, f1 score, recall, and
confusion matrix for the birdnet output and the 'perfect' scored
data.