make it kind of work
so this currently works pretty much, but the embeddings are 2
rows short. 1 can be fixed easily by adding a header to the
embeddings df, which needs to be done anyways. the other im not
entirely sure- it could be that birdnet ignores the last time
chunk because it does not hit a full 3 seconds. so I will need to
ensure the number of rows matches, and if they dont by 1, (after
adding a header to the embeddings df) then i will just delete the
bottom row of the human labeled csv, because if we cant get
embeddings for it then it doesnt matter anyways. But only needs
that fix otherwise its good. And then the script to create the
svm itself can be made