fix splitting of birdnet embeddings file
it would cut off the first 2 embeddings for each chunk in addition
to the start/stop times of the chunk. this was because the delimiter
separating the timestamps was different than the delimiter for
the embeddings themselves. now it recognizes the other delimiter
when reading in the embeddings and that fixes the issue