script to output segments detected by humans
this script takes in human labels, the wavs, and an output dir
to parse through the human labels, identify the segments in the
audio file where a human labeled a burrowing owl call, and
creates a 3 second segment wav file. Still needs to do the
same but for creating samples of the background noise with
no bird detections to create a balanced dataset with two classes
burrowing owl call and no burrowing owl call. can also be made to
create segments with specific call types rather than any
burrowing owl detection