adding pseudocode for creating dataset
starting over the script to create the dataset to employ all that
we've learned so far. Two scripts, one to run, and one to have all
the functions needed to run. create_dataset.py would be run and
would call the functions in walk_buow_labels.py (could be named
better). It would create/append to 1 folder and 1 csv all the
human labeled detections at varied lengths, and an equal number of
fixed duration noise samples from the same audio files. Now that I'm
writing this, it might be better for the noise samples to be aggregated
after all the detections have been aggregated, so we can randomly
get samples for noise from wavs that had NO detections as well. But
it's a good start.