made progress on parsing 2018 labels
the main function has been debugged, but needs another close look
through to ensure that it's not making any grave mistakes. it doesn't
crash while running through all the audio and i believe i found
a way to ensure that in the event of a duplicate file name, the
label in the all labels file will drop it out of the labels. next is
to begin the making of the segment part.