made create segments function work with subset of data
it currently will create the segments based on the filtered labels
for that audio file, create segments, and add the lines of interest
to a dataframe with the uuid of the file, the path to the file, the
call type label, the file path to the original file, and the duration
of the sample. it will return that dataframe. need to check it with
all the audio and ensure if there's any more errors, they get caught,
and make the functions to put the data from all the wavs together in 1
csv. also need to create the function that makes the noise samples.
but currently creating uuid segments and the metadata associated
seems successful.