begin adding functionality to creating noise segments
able to create an array for each second of the audio file, and a
mask in the spots that have an owl detection + a buffer. leaving
the unmasked portions as free reign for using for randomly generating
noise segments. need to add now something to randomly peruse the
unmasked portion, seeking the num of burrowing owl detections and
generating 3s samples from it, and creating a dataframe with the same
values as the create_segments function. most of the code can be
borrowed from that except for the random selection, and checking