working no_buow segment creation
this is the messiest function ive ever written but it currently
works and i want to save progress. i need to ensure it's choosing
viable start times, and there's not some weird rounding logic that
doesn't make sense. we're dealing with some values that i convert
to ints but due to the buffer added between positive segments i
think this is fine. currently makes the segments and adds it to
the other df, makes an equal number to the detection segments in
this group.