more working
added the beginnings of a script to deal with the creation of 5
folds of cleanly distrubuted data, stratified across all classes
while not splitting up clips from the same wav file to minimize
potential of leakage. also was able to get the data to all get
together in one dataframe, but still need to check the relative
start time column because something seems off about some of the
decimals. also, need to handle a none type error in the noise
segment creation because some empty dataframes are still being
passed to that one, so need to check for passing None in previous
steps. Also need to see how it will handle adding the 2017 and
2018 data together since they need to be run separately. Making the
folded dataset is separate because of this- the fact that i need to
run 2017 and 2018 separately, though i suppose it could iterate
a list and do them both at the same time? seems a little excessive
but thats the only way i could do the stratify step otherwise. i think
it will be more useful to have it as separate as i feel we'd wanna
do this in different ways and with other data.