optimized stratified group k-fold splitter
using someone elses optimization group splitting
implementation, running strat_k_folds.py uses these functions
when given the metadata file thats generated after detection and noise
segment creation. it currently doesn't append the fold number to
the metadata file, but it takes the class distributions of each
'group' in our case, groups are at the wav file level. it then
calculates the optimal split of the groups to obtain the closest
class distribution for each fold to the original class distribution.
it outputs an array with the fold number for each group, as well
as some other metrics like the actual class distribution for each fold,
because it won't be perfect given that each group is unbalanced.