create new metadata csv with the fold number in a column
this now can add the fold number assignment to the metadata but
it does create a new csv. could probably append the old one but
this is what it does. there is something funky happening with the
indexing, but this is a known issue about that creation of the
metadata.csv which would be dealth with in create_dataset pipeline
and not the fault of this script. also should create a results file
with the metrics of the actual class distributions per fold so
it can go in the readme for the dataset