handling if a wav file for a detection is less than 3s
there was one labeled wav file that was literally 1s long and 2/3
of the second was a labeled detection. it was legit and not from
another sound file. i decided to have it keep this detection and
generate a segment but it cannot generate a complementary
no_buow segment. we can change this but it seemed better to just
keep the extra detection, unless it starts happening all the time
(it wont)