add beginning of scripts to make svm for buowset
there are some differences in how buowset is structured that makes
it easier to build something custom to make the svm, as well as
the 0-padding to make the detections all 3 seconds before that.
so far, the zero padding works, BUT there are many ways to do this
differently that could minimize overfitting. This way, the padding
is added to the end of the samples. However, we should experiment
with zero padding at the beginning, both sides, and variations of
that. As well as experimenting with reducing the duration of the
no_buow segments and 0-padding that, to ensure it's not learning
the silence as a burrowing owl vocalization.

the make svm script needs to take into account the fold info in
the metadata, and change the 5 class labels to the same label type
because it will be a binary svm to start. in the future, it should
be automated to make 5 models using all the fold variations for
test and train, but this one just uses the 5th fold for test.