going to separate out the embedding extraction
going to separate the embedding extraction from the svm creation
so that one doesn't need to remake the df each time if they already
have it for their dataset. will still be able to have it remake it
for you, but could save time to separate these out and have the
make svm script only either accept the dataframe as is, or
optionally make that dataframe for you if you need it. need to change
make_svm to either handle parsing the premade data or preparing the
dataframe if you need it