decouple perch and birdnet flow
changes 2 things mainly- combines the metadata with the embed
info in the embedding file creation so that no matching needs
to occur later. this majorly speeds up the process for making the
svm itself if you already have the embed and metadata .pkl. Also
starting to decouple the make svm process for birdnet and perch, so
a parent script will need to be written for both pipelines to make
the embedding file and to make the svm, but for now both scripts
can be run separately.