create make svm script for birdnet pipeline
this script is the all-in-one script for taking birdnet embeddings
and metadata and creating the svm for you all at once. this will be
helpful for if you're adding more/changing your dataset a lot. it's
still possible to just run make_svm whether your .pkl of embedding,
fold, and label info was created from perch OR birdnet, but this script
will do the birdnet specific .pkl creation (without saving out a .pkl at
all) and make the svm for you so you don't have to juggle running
2 scripts and managing outputs and inputs.