add script to create 'loud' segments
needs some work to go through numerous sound files and give
better resulting filenames. but this script takes a sound file
and then calculates the rms across a certain number of frames you
tell it the size and hop for, giving you an array of rms values.
It then calculates the average rms, and multiplies it by 1.5
and creates 3 second segments of audio centered around the
frame that exceeded the rms threshold. it will not create
overlapping segments. there's definitely some potential issues
with this but for now seems to be able to create segments based
on the relative loudness of the whole sample. for example it
hasn't had to handle creating a segment where the values that
exceed the threshold equate for more than the specified 3s for
clip creation.