condense older versions of creating buowset segments
there were two versions i worked on, one chunked the 2017
wavs into 3s chunks and then saved them out if there was a human
buow label for that chunk. it could also create random 3s chunks
with the rest for 'no_buow' class. The other script was a bit
better because it would chunk based on the timestamps of the label,
and optionally add some extra buffer before and after in case the
duration was a bit shorter than the actual sound. now its just
the second script.