fixed line too long for my comment
leaving this script basically as-is with no linting because it
was borrowed from an other repo. I modified one line (denoted with
a comment) to account for intances where a class is defined,
but there are actually no instances of the class in the set. its
a rare error that shouldn't really happen, but during testing it
came up so it was a minor change.