60 esc extractor (#68)
* just making this into a branch

copies buowset extractor into a panda one and updated the init

* adding current scripts for esc50 instead of pandas

was working on it on my laptop, but need to commit to be able
to train on a gpu or just more powerful cpu because
my process keeps getting killed by lack of ram it seems

* lowered steps and increased epochs

the model was never evaluating because the default buowset
steps were being used (100) when esc50 is much smaller so it only had
38 steps. So it was appearing like it was not crashing and was doing
the training, but would never get evaluated. It was also not
seeming to learn fast enough with only 2 epochs (more classes than
buowset?) so epochs were increased to 10 which allowed the model
to actually start learning

* fix some pylint, still snake case stuff to fix

* fixed pylint and removed git tracking from the train.py

the train.py is always meant to just be copied and modified by
each user, so it didnt need to be tracked for esc50. the
only differences were the data extractor and the steps and epochs.
pylint done on the extractor